requirements.txt
pandas
scikit-learn
torch
mlflow

sqlalchemy==2.0.40
alembic
psycopg-binary==3.2.7
asyncpg==0.30.0
greenlet==3.2.1
psycopg2-binary

fastapi==0.115.12
uvicorn[standard]==0.34.2
python-multipart==0.0.20
h11==0.16.0
starlette==0.46.2

python-dotenv==1.1.0
annotated-types==0.7.0
anyio==4.9.0
cffi==1.17.1
click==8.1.8
dnspython==2.7.0
ecdsa==0.19.1
idna==3.10
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pytz==2025.2
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
typing-inspection==0.4.0
typing_extensions==4.13.2
